requirements.txt
google-generativeai
python-dotenv
pandas
rich
streamlit
requests
